requirements.txt
torch
transformers
accelerate
bitsandbytes
peft>=0.18.1
pillow
huggingface_hub
sentencepiece
